requirements.txt
torch>=2.1.0
torchaudio>=2.1.0
onnx>=1.15.0
onnxruntime>=1.17.0
numpy>=1.24.0
scipy>=1.11.0
soundfile>=0.12.1
pyroomacoustics>=0.7.3
pesq>=0.0.4
pystoi>=0.3.3
matplotlib>=3.7.0
tqdm>=4.65.0
pyyaml>=6.0
librosa>=0.10.0
